+++ src/Form/LegacyConsumer/templates/file.html.php
@@ -1,8 +1,9 @@
 <?php /** @var Give\Framework\FieldsAPI\File $field */ ?>
+<?php /** @var string $fieldIdAttribute */ ?>
 <input
 	type="file"
 	name="<?php echo $field->getName(); ?><?php echo $field->getAllowMultiple() ? '[]' : ''; ?>"
-	id="give-<?php echo $field->getName(); ?>"
+	id="<?php echo $fieldIdAttribute; ?>"
 	<?php echo $field->getAllowMultiple() ? 'multiple' : ''; ?>
 	<?php echo $field->isRequired() ? 'required' : ''; ?>
 	<?php echo $field->isReadOnly() ? 'readonly' : ''; ?>