+++ src/Form/LegacyConsumer/templates/html.html.php
@@ -1,4 +1,3 @@
 <?php
-
 /** @var Give\Framework\FieldsAPI\Html $field */
 echo do_shortcode( $field->getHtml() );