+++ src/PaymentGateways/resources/js/stripe-account-manager/customize-stripe-account-validation.js
@@ -8,7 +8,7 @@
 const {__, sprintf} = wp.i18n;
 
 window.addEventListener('DOMContentLoaded', function () {
-	const bodyClass = document.body.classList.contains('post.php') ? '.post.php' : '.post-new-php'
+	const bodyClass = document.body.classList.contains('post-php') ? '.post-php' : '.post-new-php'
 	const form = document.querySelector(`${bodyClass}.post-type-give_forms form[name="post"]`);
 
 	if (!form) {
@@ -16,34 +16,37 @@ window.addEventListener('DOMContentLoaded', function () {
 	}
 
 	form.addEventListener('submit', event => {
-		event.preventDefault();
+		const stripeSetting = document.querySelector('#stripe_form_account_options [name="give_stripe_per_form_accounts"]:checked');
 
-		const isCustomizeStripeAccountOptionSelected = document.querySelector('#stripe_form_account_options [name="give_stripe_per_form_accounts"]:checked').value === 'enabled';
+		// Do nothing if Stripe gateway is not active
+		if( ! stripeSetting ) {
+			return;
+		}
 
+		// Do nothing if Stripe account is not connected
+		const isCustomizeStripeAccountOptionSelected = stripeSetting.value === 'enabled';
 		if (!isCustomizeStripeAccountOptionSelected) {
-			form.submit()
 			return;
 		}
 
+		// Do nothing if default Stripe account is used
 		const hasDefaultStripeAccount = !!document.querySelector('#stripe_form_account_options .give-stripe-account-manager-list-item.give-stripe-boxshadow-option-wrap__selected');
-
-		if (!hasDefaultStripeAccount) {
-			new Give.modal.GiveNoticeAlert({
-				type: 'warning',
-				modalContent: {
-					title: __('Select Default Stripe Account', 'give'),
-					desc: __('We notice that you want to accept donation with other then global default Stripe Account but you did not selected Stripe account.', 'give'),
-				},
-			}).render();
-
-			// Open Stripe settings.
-			if ('stripe_form_account_options' !== Give.fn.getParameterByName('give_tab')) {
-				document.querySelector('a[href="#stripe_form_account_options"]').click();
-			}
-
+		if ( hasDefaultStripeAccount ) {
 			return;
 		}
 
-		form.submit()
+		event.preventDefault();
+		new Give.modal.GiveNoticeAlert({
+			 type: 'warning',
+			 modalContent: {
+				 title: __('Select Default Stripe Account', 'give'),
+				 desc: __('We notice that you want to accept donation with other then global default Stripe Account but you did not selected Stripe account.', 'give'),
+			 },
+		 }).render();
+
+		 // Open Stripe settings.
+		 if ('stripe_form_account_options' !== Give.fn.getParameterByName('give_tab')) {
+			 document.querySelector('a[href="#stripe_form_account_options"]').click();
+		 }
 	});
 });